+++ packages/pluggableWidgets/gallery-web/src/Gallery.editorPreview.tsx
@@ -1,10 +1,13 @@
 import { GUID, ObjectItem } from "mendix";
 import { createElement, ReactElement, ReactNode, useCallback } from "react";
+import { useFocusTargetController } from "@mendix/widget-plugin-grid/keyboard-navigation/useFocusTargetController";
 import { GalleryPreviewProps } from "../typings/GalleryProps";
 import { Gallery as GalleryComponent } from "./components/Gallery";
+import { useItemEventsController } from "./features/item-interaction/ItemEventsController";
+import { getPosition, useGridPositions } from "./features/useGridPositions";
+import { useClickActionHelper } from "./helpers/ClickActionHelper";
 import { useItemPreviewHelper } from "./helpers/ItemPreviewHelper";
-import { useItemSelectHelper } from "./helpers/use-item-select-helper";
-import { useItemEventsController } from "./features/item-interaction/use-item-events-controller";
+import { useItemSelectHelper } from "./helpers/useItemSelectHelper";
 
 function Preview(props: GalleryPreviewProps): ReactElement {
     const { emptyPlaceholder } = props;
@@ -13,7 +16,27 @@ function Preview(props: GalleryPreviewProps): ReactElement {
     }));
 
     const selectHelper = useItemSelectHelper(props.itemSelection, undefined);
-    const itemEventsController = useItemEventsController(selectHelper);
+
+    const { numberOfColumns, numberOfRows } = useGridPositions({
+        phoneItems: props.phoneItems ?? 1,
+        tabletItems: props.tabletItems ?? 1,
+        desktopItems: props.desktopItems ?? 1,
+        totalItems: items.length
+    });
+    const getPositionCallback = useCallback(
+        (index: number) => getPosition(numberOfColumns, items.length, index),
+        [numberOfColumns, items.length]
+    );
+
+    const focusController = useFocusTargetController({
+        rows: numberOfRows,
+        columns: numberOfColumns,
+        pageSize: props.pageSize ?? 0
+    });
+
+    const clickActionHelper = useClickActionHelper({ onClick: props.onClick });
+
+    const itemEventsController = useItemEventsController(selectHelper, clickActionHelper, focusController);
 
     return (
         <GalleryComponent
@@ -49,6 +72,8 @@ function Preview(props: GalleryPreviewProps): ReactElement {
             tabletItems={props.tabletItems!}
             selectHelper={selectHelper}
             itemEventsController={itemEventsController}
+            focusController={focusController}
+            getPosition={getPositionCallback}
         />
     );
 }