+++ packages/pluggableWidgets/gallery-web/src/helpers/SelectActionHelper.ts
@@ -1,3 +1,3 @@
 import { SelectActionHandler } from "@mendix/widget-plugin-grid/selection";
 
-export class ItemSelectHelper extends SelectActionHandler {}
+export class SelectActionHelper extends SelectActionHandler {}